Complex query fix, cascade delete

+++ src/main/java/cz/cvut/fit/tjv/moment/business/AbstractCrudService.java
@@ -59,6 +59,11 @@ public abstract class AbstractCrudService<K, E, REPOSITORY extends JpaRepository
     }
 
     public void deleteById(K id) {
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        }
+        else {
+            throw new NoSuchElementException("No such element to update.");
+        }
     }
 }
